+++ test/fixtures/schema.sql
@@ -1083,6 +1083,11 @@ create function test.many_inout_params(INOUT num int, INOUT str text, INOUT b bo
   select num, str, b;
 $$ language sql;
 
+CREATE FUNCTION test.variadic_param(VARIADIC v TEXT[] DEFAULT '{}') RETURNS text[]
+LANGUAGE SQL AS $_$
+  SELECT v
+$_$;
+
 create or replace function test.raise_pt402() returns void as $$
 begin
   raise sqlstate 'PT402' using message = 'Payment Required',